feat/CLI tool for monitoring warp route balances (#3204)
### Description

Motivation

After Hyperlane launches a warp route, we need to monitor the balances
for the grafana dashboard.

Before

1. Ask Kunal to run the script  
2. Kunal will ask Rossy/Yorke for the warp route configs (tokenAddress,
hyperc20Address, etc)
3. Kunal will run it locally and then deploy it in on k8s

After

1. Make a config yaml file to test locally and deploy on k8s with new
workload name as the file.


PR Details

- Replaced the statically typed warp route configs with a more concise
zod config which parses the yaml file.
- Added the yaml files for nautilus, nexus and injective warp routes.
- Note: The ZOD config is compatible with the updated registry schema
but incompatible with the WarpRouteConfigSchema mostly due to the
different toplevel keys (base and synthetic vs chainNames) and hence the
duplication.
- The chainNames is preferred to base and synthetic bc of its uniqueness
as a key and consistency with the rest of the tooling. (eg. injective
and inevm has native types on both side where it's not obvious which
should be base or synthetic)
- Note: The file names need to follow the `${route}-deployments.yaml`
still to be properly parsed by the deploy-warp-monitor.ts script (this
path should be relative to monorepo root)


### Drive-by changes

- Warp route monitor can now fetch native balances through the
CosmNativeTokenAdapter.

### Related issues

None

### Backward compatibility

No, the script needs a new arg

### Testing

Manual