feat: fetch chain-specific metrics in standalone task (#3214)
### Description

Spawning a tokio task to fetch chain-specific metrics in provider
middleware was wasteful because we ended up with 15 tasks doing the same
job. Now there is a single task that is spawned inside the relayer,
which I added in the struct that used to only fetch relayer balance.

Some questions:
- I'm realising in the current state I'm probably removing these metrics
from validators. Lmk if yes and I'll add back
- the chain-specific metrics were moved out of `MiddlewareMetrics`
because they're no longer part of middleware, but I wasn't sure where to
place them. Maybe `CoreMetrics` is a better place than the current
`custom_metrics.rs` file?

`tokio-metrics` turned out not to be useful because we'd need to
instrument every call site of `tokio::spawn` with it. We should still do
it but as a separate task imo.

### Drive-by changes

- This PR also makes it very easy to add the metrics tasks for new
chains, by extending the `HyperlaneProvider` trait with a
`get_chain_metrics` call which seems reasonably general to me to have
implemented by all providers. (Bc we currently only track these for evm
chains)
- the description, naming and logic of the `gas_price` metric was also
changed to support new chains

### Related issues

- Fixes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/3047

### Backward compatibility

Yes

### Testing

Still need to manually test