Hacky solana balance monitoring (#3432)
### Description

- See https://github.com/hyperlane-xyz/issues/issues/1047
- Intentionally going with a "hacky" short term approach because we
consistently hit this, we're unlikely to prioritize a better solution
for v3, and this change is only for v2 which we are in maintenance-only
mode, so it's ok to hardcode these accounts

### Drive-by changes

- New monorepo base image to fix building, which requires node 18
apparently

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

<!--
What kind of testing have these changes undergone?

None/Manual/Unit Tests
-->