feat: add HypERC20 checker (#3517)
### Description

- Added HypERC20App (extends GasRouterApp) and HypERC20Checker (extends
RouterChecker bc there's no GasRouterChecker - link to issue)
- remoteChains fetches the remote chains from the local router address
is it given. This means, the abstract routerAddress in multiGeneric had
to be made and this call had to awaited in a few places where it's used.
**Why?** previously we assumed that the routers will be fully-connected
and will match 1:1 with the config. This is too strong of an assumption
for ICAs given we inherit the config through `getRouterConfig` and the
ICA accounts will only need to exist on chains which don't have safes.
Plus, in some cases, the chains implied by the app factories vs checker
config don't match and this leads to further issues down the line while
checking for mailbox client properties or in the govern. Checking the
onchain enrollments is the most safe and resilient approach we can take
especially given that having a checker for every RouterApp isn't
feasible.

### Drive-by changes

- adding native keys to warp artifacts for the appHelper to work in
App.fromAddressMap. Note: if there's something using the router key,
this isn't breaking unless you expect only one key.

### Related issues

- fixes https://github.com/hyperlane-xyz/issues/issues/1191

### Backward compatibility

Yes

### Testing

Manual

---------

Signed-off-by: Paul Balaji <paul@hyperlane.xyz>
Co-authored-by: Paul Balaji <paul@hyperlane.xyz>
Co-authored-by: J M Rossy <jm.rossy@gmail.com>
Co-authored-by: Trevor Porter <trkporter@ucdavis.edu>
Co-authored-by: Daniel Savu <23065004+daniel-savu@users.noreply.github.com>