Allow protocol-specific indexing (#3544)
### Description

- Generalizes contract syncer hashmaps to store trait objects instead of
concrete structs, since both `SequencedDataContractSync` and
`WatermarkContractSync` implement `ContractSyncer`
- The nice thing this enables is being able to instantiate a different
type of syncer for the same contract, for different chain protocols
(e.g. for IGP, we can use a sequenced syncer on sealevel and cosmwasm,
and a watermark syncer on EVM).
- The problem is that `ContractSyncer` is implemented on different trait
bounds on `SequencedDataContractSync` and `WatermarkContractSync`. Most
notably, the `SequencedDataContractSync` impl bounds the indexed type by
the `Sequenced` trait. This means that if a type is to be indexed by
more than one syncer strategy, it needs to implement the union of bounds
of those strategies
- **Blocker:** In the IGP, this means that `InterchainGasPayment` needs
to implement `Sequenced` and we have no obvious way of getting that just
from the type. Even if the sequence is derived from the type, it has to
be a sensible value, since indexing reliability depends on it.
- Similar to the point above, we now have `contract_syncs`, a single
function that instantiates contract syncs for all chains. Since
`contract_syncs` must be able to build a `ContractSyncer` trait obj for
all syncer types, it bounds both `T` (the indexed type) and `D` (the log
store type) with a union of bounds, so it may be too strict for some
types in the future.
- The scraper now uses forward-backward indexing instead of just forward
indexing, because that simplified the refactor - there's now a
convenience trait called `TryFromWithMetrics` that allows adding this
bound: `SequenceIndexer<T>: TryFromWithMetrics<ChainConf>,`. This isn't
really extendable if we use non-sequence indexers, but it's fine for now
since this is used everywhere
- replaces the macro-based contract syncer builder with generic fns

Still todo:
- add new `IndexingDecorator<T>` struct that can wrap
`InterchainGasPayment`, so we can use sequence-aware indexing for the
IGP
- remove sequence logic from the watermark cursor

### Drive-by changes

<!--
Are there any minor or drive-by changes also included?
-->

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

<!--
What kind of testing have these changes undergone?

None/Manual/Unit Tests
-->