Update IGP gas oracles with new heuristics, get Cosmos gas prices from registry (#3662)
### Description

The original and main motivation of this PR was to feel confident when
running `deploy.ts -m igp` off of main. Prior to this, running it off
main could result in misconfiguring gas oracles, particularly when it
came to non-EVM remotes. See
https://github.com/hyperlane-xyz/issues/issues/1178 for some context.

This PR does a number of things:
- Fetches Cosmos gas prices from the Cosmos registry
- the `gasPrices.json` now lets you specify the decimals relating to the
amount in there. There are 2 reasons for this: it's easier for us to
deal with gwei values for EVM chains (as we often end up manually
changing these instead of just relying on the script). The second reason
is covered in the next bullet
- Handles decimals when it comes to gas prices. The existing tooling
assumed gas prices would always be denominated in wei. An extra weird
situation arises with non-EVM chains: chains like Neutron have gas
prices that are quotes in the smallest denomination (ie the equivalent
of "wei" for NTRN), but the gas price is not an integer and is very
small, e.g. `0.0053`. This poses issues because the IGP contracts
require the gas price that's configured onchain to be an integer. Simply
rounding up from 0.0053 to 1 results in almost 2 orders of magnitude
difference :(. So the solution here is to have some logic that
accommodates really tiny gas prices by scaling up the configured gas
price and scaling down the configured token exchange rate.
- Applies some "non-linear" gas pricing heuristics. See
https://github.com/hyperlane-xyz/issues/issues/1204 and
https://discord.com/channels/935678348330434570/935678739663192184/1232293417309044818
for some context
- Running `update-agent-config.ts` will fetch gas prices for Cosmos
chains from the Cosmos registry and ensure they're accurate

### Drive-by changes

Some logging messages when applying IGP gas oracle changes to make it
more clear what the change actually does

### Related issues

- Fixes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/3386
- Fixes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/3484
- Fixes https://github.com/hyperlane-xyz/issues/issues/1178
- Fixes https://github.com/hyperlane-xyz/issues/issues/1206
- Fixes https://github.com/hyperlane-xyz/issues/issues/1204

### Backward compatibility

Fully backward compatible

### Testing

Ran deploy.ts for the IGP on all chains!

---------

Co-authored-by: nambrot <nambrot@googlemail.com>