fix: use secret RPC urls in infra, apply environment-specific registry overrides (#3875)
### Description

A while back, it seems secret RPC URLs stopped being used by monorepo
services (see
https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/3179/files#diff-ba3b7bea2b5c046acd88d25ed382df875b49d77e061f2ef8347ffcc741c0353eR34).
This adds them back.

Some things to note on the approach:
- Sadly, due to the way some of the infra config generation works, we're
forced to use synchronous interfaces for getting chain metadata in a
bunch of places. This is why `getRegistry` returns a
`FileSystemRegistry`, which is a `SynchronousRegistry`.
- Applying secrets / overrides requires us to do things async for two
reasons -- one is just that fetching from GCP secrets is an async
operation, the other is that a `MergedRegistry` is not a
`SynchronousRegistry`. The approach I've taken is that there are two
ways of getting a registry - using `getRegistry` will give the
synchronous version that can continue to be used by our sync config
generation, but it won't have any environment-specific overrides (tx
overrides or secrets). To get an async registry with
environment-specific overrides, the `getRegistry` function on the
EnvironmentConfig is used.
- Overrides are now more succinct as the merging happens later in the
pipeline

### Drive-by changes

- Started moving away from using the gcloud CLI under the hood for GCP
operations and toward using the GCP secret SDK
- Updated the mainnet3 key funder to fix an incident, but will follow up
with a fresh deploy once this is mergd

### Related issues

- These are the code changes for
https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/3774, but
I'll follow up with fresh deploys for the monorepo services

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

<!--
What kind of testing have these changes undergone?

None/Manual/Unit Tests
-->

---------

Co-authored-by: Connor McEwen <connor.mcewen@gmail.com>
Co-authored-by: Yorke Rhodes <yorke@hyperlane.xyz>
Co-authored-by: J M Rossy <jm.rossy@gmail.com>