feat: batching igp (#3870)
### Description

Deducts IGP cost from senders of messages within a batch, based on the
total gas used by the batch and the estimates for submitting messages
individually.

The formula is `gas_used_by_operation = gas_used_by_tx *
(operation_estimated_gas / total_estimated_cost)`

Note that the individual estimate have a buffer value added to them
(currently 75k), which will slightly skew proportions, though by a
negligible amount.

For example, given these estimates in a batch: 150k, 160k, 180k, if we
add 80k to each, we go from e.g. 0.312 for the first message to 0.319 ->
2.25% error which isn't so bad. The error can be larger if operations
within a batch have very different estimates, but realistically within a
5% range based on back-of-the-napkin calculations.

### Drive-by changes

- The batching feature introduced an bug whereby gas expenditure would
only be deducted in the `confirm` step. Now this is done in the `submit`
step to account for txs that revert
- Sealevel e2e can now be disabled for faster iteration! I've not done
the cleanest job, but even the fact that we have this will reduce the
time to run e2e locally by more than half. To use it, set
`SEALEVEL_ENABLED=false` when running `run-locally`
- e2e uses a new utility called `get_matching_lines`, that can be used
to count (and in the future _parse_) logs, to reconstruct the state of
the relayer and have more expressive correctness checks. This is used to
make sure that gas is deducted for all messages, including those in
batches.

### Related issues

- Fixes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/3709

### Backward compatibility

Yes

### Testing

E2E, using `get_matching_lines`