feat: relayer queue operation labels (#4003)
### Description

- adds status labels to operations in the relayer queues. In practice,
the most useful ones are those for when a message is retried due to an
error. Labels are defined as enums to keep the cardinality finite, the
size of prometheus metrics could get very large
- whenever a message is pushed to a queue, the interface requires
specifying the new status as a `Option<PendingOperationStatus>`. If
`None`, the status is left as before (such as in the case when an
operation isn't ready to be retried).
- a limitation is that labels aren't stored in the DB, so messages are
loaded with the `FirstPrepareAttempt` label. I just triggered retries
for destinations I wanted to get labels for, but we could persist to the
db in a follow-up PR
- Example prep queue labels for inevm:
<img width="621" alt="Screenshot 2024-06-21 at 18 05 58"
src="https://github.com/hyperlane-xyz/hyperlane-monorepo/assets/23065004/f53b2858-371f-4b6b-86d9-5c21409d641f">


### Drive-by changes

- removes `make_op_try` as it doesn't really add anything that couldn't
be done with regular Rust
- bumps the `strum` version because there was a bug in the previous one,
preventing inner enum fields from being converted to strings

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

<!--
Yes

### Testing

Manual