feat: Report git commit hash as agent metadata (#4282)
### Description

* We are storing git commit hash as an environment variable when we
compile the source code of an agent.
* Validator uses git commit hash as a agent version in
`metadata_latest.json` file.
* File `metadata_latest.json` is overwritten every time when validator
is restarted.
* Git commit hash is extracted by `git` util, so, it should be available
during build time.
* If the directory with the source code does not contain `.git` folder,
git commit hash will be defaulted to `VERGEN_IDEMPOTENT_OUTPUT`.

Agent Metadata is propagated to Relayer and Scraper, but is not used
there. We can use it in the future, reporting git hash commit in the
logs, for example.

### Backward compatibility

Yes

### Testing

* Tested locally with E2E Tests.
* Tested against S3 by deploying docker image into k8s

---------

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>