feat: deploy pzETH PoC, various tooling tweaks and fixes, and fix an edge case in the multisig ISM program (#4359)
### Description

A few small tweaks:
- This now transfers the metadata pointer and metadata update
authorities. e.g. on this spl-token I created, this is what it looks
like, where there are distinct authorities for the metdata pointer and
metadata extensions:
```
$ spl-token display 5kFYTUUG5N4uoxAEFLBB5n45eAzifAGKwcY1uoogn6SS -u mainnet-beta

SPL Token Mint
  Address: 5kFYTUUG5N4uoxAEFLBB5n45eAzifAGKwcY1uoogn6SS
  Program: TokenzQdBNbLqP5VEhdkAS6EPFLC1PHnBqCXEpPxuEb
  Supply: 0
  Decimals: 9
  Mint authority: 5kFYTUUG5N4uoxAEFLBB5n45eAzifAGKwcY1uoogn6SS
  Freeze authority: (not set)
Extensions
  Metadata Pointer:
    Authority: 9bRSUPjfS3xS6n5EfkJzHFTRDa4AHLda8BU2pP4HoWnf
    Metadata address: 5kFYTUUG5N4uoxAEFLBB5n45eAzifAGKwcY1uoogn6SS
  Metadata:
    Update Authority: 6DjHX6Ezjpq3zZMZ8KsqyoFYo1zPSDoiZmLLkxD4xKXS # <-- this is because this was my default keypair in my config
    Mint: 5kFYTUUG5N4uoxAEFLBB5n45eAzifAGKwcY1uoogn6SS
    Name: Renzo Restaked LST
    Symbol: pzETH
    URI: https://raw.githubusercontent.com/hyperlane-xyz/hyperlane-registry/12660fd34d30e960a748d87408a8d88f634f7454/deployments/warp_routes/pzETH/ethereum-solana-metadata.json
```

So I added a step to transfer these away as well. I think it's possible
to maybe initialize it right off the bat to the mint, but for now this
works

- Made `adjust_gas_price_if_needed` idempotent - ran into an issue where
this was called twice before a tx was submitted because of the warp
route edge case where there's just a single SVM based chain
- my CLI config has the devnet deployer key
`6DjHX6Ezjpq3zZMZ8KsqyoFYo1zPSDoiZmLLkxD4xKXS` set as the default
instead of the mainnet deployer
`9bRSUPjfS3xS6n5EfkJzHFTRDa4AHLda8BU2pP4HoWnf`. This caused some issues
where some CLI commands assumed that the default key was the mainnet
deployer. Added some flags to some commands to ensure this is correct
- Updated the Solana gas price to be accurate - it was charging
thousands with the provided gas price - it was off by a few orders of
magnitude

And one major fix:
- As part of our v3 update, a subtle ISM bug was introduced -- the index
of the signed checkpoint was being set to the message's nonce, when it
should've been from the metadata's merkle_index. This poses issues if
the message nonce != the checkpoint index, which is what happened when I
sent a test message from Ethereum to Solana. Yikes! I redeployed a new
ISM program and set it as the default on the Mailbox, and all was fixed.
I updated the tests to specifically cover this case

### Drive-by changes

<!--
Are there any minor or drive-by changes also included?
-->

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

<!--
What kind of testing have these changes undergone?

None/Manual/Unit Tests
-->

---------

Co-authored-by: Daniel Savu <23065004+daniel-savu@users.noreply.github.com>