fix: increase wait time for postgres to be ready in cosmos e2e (#4411)
## fix: increase wait time for postgres to be ready in cosmos e2e

increase time waiting for postgres to be ready, to reduce this sort of
flake
https://github.com/hyperlane-xyz/hyperlane-monorepo/actions/runs/10665491633/job/29559169324?pr=4410#step:13:2998

based on the logs it looks like 10s is just on the border, adding an
extra 5s to give headroom


## drive-by fix: disable fail-fast on e2e-matrix

currently only the non-cosmwasm tests are required to merge, therefore
we should not fail the entire matrix if the cosmos job fails. This also
means we can test that the issues we're seeing with the `cosmwasm` job
do not affect the `non-cosmwasm` one.