fix: gracefully handle missing objects for GCS checkpoints (#4453)
### Description

The issue with the validator was that it wasn't handling the bucket's
404 response as a correct response when the object is not found when
fetching a specific checkpoint or latest index, thus it was getting
stuck in a loop failing to submit a checkpoint or get the latest stored
index.

I fixed it by catching the 404 response in `gcs_storage.rs` and
returning `Ok(None)` instead of raising an error, the same way it's
handled in S3 Storage.

### Drive-by changes

Bumped ya-gcp to the latest 0.11.3

### Related issues

- Fixes #4451 

### Backward compatibility

Yes

### Testing

Manual