fix: GCS correctly splitting storage location into bucket and folder (if present) (#4539)
### Description

Previously Checkpoint Syncer wasn't correctly parsing the storage
location string. If no folder was provided, it treated the object name
as folder, e.g. reading `gs://bucket-name/gcsLatestIndexKey` initialized
StorageClient with bucket `bucket-name` and folder `gcsLatestIndexKey`.
This was due to incorrect logic for splitting the string by `/` that
expected a folder to be present.

In this PR I fixed this logic, treating a location with only one `/` as
_bucket/object_ and with two `/`s as _bucket/folder/object_.

### Backward compatibility

Yes

### Testing

Manual

---------

Co-authored-by: Trevor Porter <tkporter4@gmail.com>