feat: HyperlaneDb trait (#4609)
### Description

Defines a `HyperlaneDb` trait for use in the validator submitter, so a
mock can be passed in to simulate indexed data from reorged state.

Although the `HyperlaneDb` trait can't be `Clone` to be object safe, as
long as we use `Arc`s everywhere, all will be ok. Eventually the aim is
to replace all direct usage of `HyperlaneRocksDb` with `Arc<dyn
HyperlaneDb>`.