feat: Generate Cosmos address for Injective (#4618)
### Description

Injective uses different logic compared to Neutron and Osmosis to
generate account addresses from public key. It is implemented in this PR
so that sender address is correctly generated from signer info of
transaction.

Depending on the type of public key communicated in the transaction, we
shall generate either Bitcoin-like or Ethereum-like address for
Injective.

We can also specify the type of account id for the signer. Bitcoin-like
will be chosen by default, so, we don't need to change the configuration
of Relayer.

### Drive-by changes

Made some errors to be warnings. We'll add metrics to measure how
frequent the issues are and decide how to fix them:
1. transaction contains multiple contract execution messages
2. transaction contains fees in unsupported denominations

Made Scraper to refrain from recording transaction which it cannot parse
properly.

### Backward compatibility

Yes

### Testing

Added unit tests
Manual testing with Injective blockchain

---------

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>