fix: update HyperlaneDomainTechnicalStack with latest stacks (#4688)
### Description

fix: update HyperlaneDomainTechnicalStack with latest stacks

We added new technical stacks, and updated these in registry, but did
not update the enum in the agents. this caused a hard failure to load
the configs when we tried deploying to prod.

Note: we're omiting `zksync` because that is soon to be moved from a
"tech stack" into a "protocol type", and we have no chains in registry
with a `zksync` technical stack configured yet anyway.

### Drive-by changes

update existing list of mainnets with "other" technical stack. can't
just remove them right now because we have to ensure every value in the
enum is mapped

### Related issues

na

### Backward compatibility

yes

### Testing

todo: fix e2e