fix(e2e): reenable gas enforcement check in e2e (non-cosmwasm) (#5105)
### Description

- Move to valid 'gaspaymentenforcement' for e2e, which was earlier
shortcircuiting to none (the default value for enforcement policy). This
meant we were not checking if the relayer was enforcing a non-zero gas
payment for either sealevel or EVM gas payments.
- The igpHook set as part of the default hook for Kathy was not
configured properly, hence was giving a quote of 0. Now, Kathy does set
the IGP config (oracle and overhead). (Note: Kathy is soon to be
deprecated, so this is more of a quick patch for e2e coverage).
- Added a check for deserilization of 'gaspaymentenforcement' and panic
if invalid to prevent this bug in the future.



### Drive-by changes

<!--
Are there any minor or drive-by changes also included?
-->

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

<!--
What kind of testing have these changes undergone?

None/Manual/Unit Tests
-->