chore: retryable errors should still be logged as warnings (#5184)
### Description

Retryable errors often indicate that something abnormal is happening
(such as we falling back to other providers, which may be causing
issues). It's much easier to notice this if the logs are emitted are
`warn` level. Non-retryable errors should be logged at `error` level to
highlight the relative severity to retryable ones