fix: zksync default build artifact (#5269)
### Description

The aim of this PR is to satisfy the exports in `package.json` and to
ensure that the SDK build command succeeds in the upcoming ZKSync PRs
(imported by ContractVerification classes). In the `dist` folder, this
file is overwritten with the correct values when 'build:zk' is run.

### Drive-by changes

- Default export of buildArtifact as an empty object

### Related issues

None

### Backward compatibility

Yes

### Testing

Build passing in the upcoming ZKSync PRs

---------

Co-authored-by: Morteza Shojaei <31728528+mortezashojaei@users.noreply.github.com>
Co-authored-by: mshojaei-txfusion <138107084+mshojaei-txfusion@users.noreply.github.com>