feat(relayer): don't retry old messages (#5455)
### Description

We've recently noticed the submitter attempts to process many old
messages at the same time, because they have very similar
`next_retry_attempt`s. These old messages are unlikely to be
processable, and in fact starve new "healthy" messages, sometimes
leading to lowered throughput.

There was an attempt to make this less of an issue in
https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/5416, but we
decided old messages are not worth retrying at all given we offer the
ability to deliver messages on-demand in the hyperlane CLI.

Not retrying old messages means we don't even have to push them to the
submitter when loaded from the db - we just skip them.

This PR:
- introduces a `HYP_MAXMESSAGERETRIES` env var which defaults to 66 if
not set. The default of 66 is picked under the assumption that the first
48 retries take about 1 day
([formula](https://github.com/hyperlane-xyz/hyperlane-monorepo/blob/4d6fd2f8d72dbc882572e73ba599b80870f2bf67/rust/main/agents/relayer/src/msg/pending_message.rs#L674)).
The remaining 22 retries will take at least `21 * 22 / 2` hours (see the
formula), which is about 11 days. In total that is almost 2 weeks, which
@nambrot confirmed is how long we should retry for. If the default is
incorrect we can set a higher env var to override. I also manually
checked the queues for the arbitrum and base destinations, and there was
no message with a retry count of 100 but most old ones had more than 70
attempts.
- skips messages read from the db whose retry count is >=
MAX_MESSAGE_RETRIES
- for messages already loaded in the submitter, sets
`Duration::from_secs(u32::MAX)` as the next retry attempt, which is far
enough into the future as to not be attempted at all

Prepare queues are expected to get much smaller because of the skipping,
which hopefully also lowers memory pressure.

### Drive-by changes

- Makes the db field in `MessageContext` a trait obj for easier future
mocking

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

Yes - if `HYP_MAXMESSAGERETRIES` is set to a very high value, it's as if
this feature is disabled

### Testing

Unit testing of the message backoff logic and of inner logic in
`maybe_from_persisted_retries`. The env var loading isn't tested but is
very similar to how we load env vars
[elsewhere](https://github.com/hyperlane-xyz/hyperlane-monorepo/blob/4d6fd2f8d72dbc882572e73ba599b80870f2bf67/rust/main/chains/hyperlane-sealevel/src/rpc/client.rs#L412)
- longer term it should be turned into a proper agent config (the other
ones too).