feat: Expand warp deploy config in warp check (#5483)
### Description

When using `WarpDeployConfig` we often assume certain defaults, however
when we compare against `warp read` config, it includes all the
"expanded" config. This PR adds "config expansion", so that checking
actually can be done. As part of that, it removes the practice of `warp
read` after deploy in the warp check e2e tests so that the actual
comparison against the intended deploy config can happen (instead of the
inferred config from warp read).

This is related to
https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/5358 and in
fact lifts some of the convenience function into `configUtils.ts` fyi
@mshojaei-txfusion

### Drive-by changes

The warp-check e2e test setup changes as mentioned

### Backward compatibility

Yes

### Testing

Local e2e tests and manual