feat: Clarify Reprepare reason for Hyperlane Application Context (#5366)
### Description

Clarify Reprepare reason for Hyperlane Application Context

If application context indicates that a Hyperlane message was being
delivered as part of warp route, we make more verifications (depending
on the vm) so that we can differentiate between error of estimating gas
for the delivery and messages which does not conform the rules applied
to a warp route. We change the reprepare reason for non-conforming
messages so that we can differentiate such events in metrics.

### Drive-by changes

- Minor refactoring

### Related issues

- Fixes https://github.com/hyperlane-xyz/issues/issues/1414

### Backward compatibility

Yes

### Testing

Manual since application specific checks are not part of CI

---------

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>