fix: More exhaustive search for a relevant transaction in Solana block for Scraper (#5558)
### Description

Currently, Scraper cannot make progress (either forward or backward) on
Solana chain if it encounters a transaction which contains two separate
entries of Mailbox program acting on two different Hyperlane messages.
If Scraper encounters such a transaction on the forward indexing, it
stalls and new dispatched messages are not being indexed anymore. If
Scraper encounters such a transaction on the backward indexing, Scraper
will not index historical dispatched messages.

The root cause of the issue is that Scraper encounters the first Mailbox
program entry in the transaction, but since this entry does not act an a
given Hyperlane message, the whole transaction is considered to be
irrelevant to the given Hyperlane message.

This change fixes the issue: Scraper will continue iterating over the
list of instructions (programs) included into a transaction until it
exhaust all of them, so that it does not miss the entry of Mailbox
program which acts on the Hyperlane message in question.

### Backward compatibility

Yes

### Testing

Unit test

---------

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>